# self_aware/openapi/agent-gateway.json
{
  "components" : {
    "schemas" : {
      "AgentErrorResponse" : {
        "properties" : {
          "candidates" : {
            "items" : {
              "$ref" : "#/components/schemas/EntityCandidate"
            },
            "type" : "array"
          },
          "code" : {
            "type" : "string"
          },
          "confirmed_count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "current_count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "hint" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "param" : {
            "type" : "string"
          },
          "precondition" : {
            "type" : "string"
          },
          "step" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "code", "message" ],
        "type" : "object"
      },
      "AgentMetadataResponse" : {
        "properties" : {
          "capabilities" : {
            "items" : {
              "$ref" : "#/components/schemas/CapabilityMetadata"
            },
            "type" : "array"
          }
        },
        "required" : [ "capabilities" ],
        "type" : "object"
      },
      "BlastRadius" : {
        "enum" : [ "none", "single", "group", "branch", "organisation" ],
        "type" : "string"
      },
      "CapabilityMetadata" : {
        "properties" : {
          "blast_radius" : {
            "$ref" : "#/components/schemas/BlastRadius"
          },
          "description" : {
            "type" : "string"
          },
          "disambiguate_from" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "effect" : {
            "$ref" : "#/components/schemas/EffectMetadata"
          },
          "id" : {
            "type" : "string"
          },
          "module" : {
            "type" : "string"
          },
          "params" : {
            "items" : {
              "$ref" : "#/components/schemas/ParamMetadata"
            },
            "type" : "array"
          },
          "preconditions" : {
            "items" : {
              "$ref" : "#/components/schemas/PreconditionMetadata"
            },
            "type" : "array"
          },
          "read_only" : {
            "type" : "boolean"
          },
          "reverses" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        },
        "required" : [ "blast_radius", "description", "disambiguate_from", "effect", "id", "module", "params", "preconditions", "read_only", "version" ],
        "type" : "object"
      },
      "CapabilityVersion" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "version" ],
        "type" : "object"
      },
      "CapabilityVersionsResponse" : {
        "properties" : {
          "versions" : {
            "items" : {
              "$ref" : "#/components/schemas/CapabilityVersion"
            },
            "type" : "array"
          }
        },
        "required" : [ "versions" ],
        "type" : "object"
      },
      "EffectMetadata" : {
        "properties" : {
          "confirmation_template" : {
            "type" : "string"
          },
          "creates" : {
            "type" : "string"
          },
          "facts" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "notifies" : {
            "type" : "string"
          },
          "pending_template" : {
            "type" : "string"
          },
          "reply_template" : {
            "type" : "string"
          }
        },
        "required" : [ "facts", "reply_template" ],
        "type" : "object"
      },
      "EntityCandidate" : {
        "properties" : {
          "context" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "label" ],
        "type" : "object"
      },
      "ExecuteOutcome" : {
        "enum" : [ "completed", "partial", "failed" ],
        "type" : "string"
      },
      "ExecuteRequest" : {
        "properties" : {
          "plan" : {
            "$ref" : "#/components/schemas/Plan"
          },
          "sentence" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "plan", "sentence", "token" ],
        "type" : "object"
      },
      "ExecuteResponse" : {
        "properties" : {
          "outcome" : {
            "$ref" : "#/components/schemas/ExecuteOutcome"
          },
          "plan_id" : {
            "type" : "string"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/ExecutedStep"
            },
            "type" : "array"
          }
        },
        "required" : [ "outcome", "plan_id", "steps" ],
        "type" : "object"
      },
      "ExecutedStep" : {
        "properties" : {
          "capability_id" : {
            "type" : "string"
          },
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "data" : { },
          "error" : {
            "$ref" : "#/components/schemas/AgentErrorResponse"
          },
          "reply" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/StepStatus"
          },
          "step" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "capability_id", "status", "step" ],
        "type" : "object"
      },
      "ParamMetadata" : {
        "properties" : {
          "allowed" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "default_value" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "meaning" : {
            "type" : "string"
          },
          "multiple" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "required" : {
            "type" : "boolean"
          },
          "resolver" : {
            "type" : "string"
          },
          "type" : {
            "$ref" : "#/components/schemas/ParamType"
          }
        },
        "required" : [ "allowed", "meaning", "multiple", "name", "required", "type" ],
        "type" : "object"
      },
      "ParamType" : {
        "enum" : [ "string", "integer", "decimal", "boolean", "date" ],
        "type" : "string"
      },
      "ParamValue" : {
        "properties" : {
          "chosen_id" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "from_step" : {
            "format" : "int32",
            "type" : "integer"
          },
          "raw" : {
            "type" : "string"
          },
          "value" : { }
        },
        "type" : "object"
      },
      "Plan" : {
        "properties" : {
          "plan_id" : {
            "type" : "string"
          },
          "session_id" : {
            "type" : "string"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/PlanStep"
            },
            "type" : "array"
          }
        },
        "required" : [ "plan_id", "session_id", "steps" ],
        "type" : "object"
      },
      "PlanStep" : {
        "properties" : {
          "capability_id" : {
            "type" : "string"
          },
          "capability_version" : {
            "type" : "string"
          },
          "params" : {
            "additionalProperties" : {
              "$ref" : "#/components/schemas/ParamValue"
            },
            "type" : "object"
          },
          "step" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "capability_id", "capability_version", "params", "step" ],
        "type" : "object"
      },
      "PreconditionMetadata" : {
        "properties" : {
          "hint" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "text" : {
            "type" : "string"
          }
        },
        "required" : [ "hint", "id", "text" ],
        "type" : "object"
      },
      "PreflightRequest" : {
        "properties" : {
          "plan" : {
            "$ref" : "#/components/schemas/Plan"
          }
        },
        "required" : [ "plan" ],
        "type" : "object"
      },
      "PreflightResponse" : {
        "properties" : {
          "confirmation" : {
            "type" : "string"
          },
          "expires_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "plan_id" : {
            "type" : "string"
          },
          "requires_confirmation" : {
            "type" : "boolean"
          },
          "steps" : {
            "items" : {
              "$ref" : "#/components/schemas/PreflightStepResult"
            },
            "type" : "array"
          },
          "token" : {
            "type" : "string"
          },
          "warnings" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "expires_at", "plan_id", "requires_confirmation", "steps", "token", "warnings" ],
        "type" : "object"
      },
      "PreflightStepResult" : {
        "properties" : {
          "capability_id" : {
            "type" : "string"
          },
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "line" : {
            "type" : "string"
          },
          "pending" : {
            "type" : "boolean"
          },
          "resolved" : {
            "items" : {
              "$ref" : "#/components/schemas/ResolvedEntity"
            },
            "type" : "array"
          },
          "step" : {
            "format" : "int32",
            "type" : "integer"
          },
          "unit" : {
            "type" : "string"
          }
        },
        "required" : [ "capability_id", "pending", "resolved", "step" ],
        "type" : "object"
      },
      "ResolvedEntity" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "param" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "label", "param" ],
        "type" : "object"
      },
      "SessionCapabilitiesResponse" : {
        "properties" : {
          "capability_ids" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "user_id" : {
            "type" : "string"
          }
        },
        "required" : [ "capability_ids", "user_id" ],
        "type" : "object"
      },
      "StepStatus" : {
        "enum" : [ "succeeded", "replayed", "failed", "not_run" ],
        "type" : "string"
      }
    }
  },
  "info" : {
    "description" : "The only endpoints the AI layer calls. Plans name capability ids, never URLs.",
    "title" : "School backend: agent gateway",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/agent/execute" : {
      "post" : {
        "operationId" : "execute",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExecuteResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "INVALID_PLAN"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "UNAUTHENTICATED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "TOKEN_INVALID, NOT_PERMITTED"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "TOKEN_EXPIRED, STALE_VERSION"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "NOT_IMPLEMENTED"
          }
        },
        "tags" : [ "execute-controller" ]
      }
    },
    "/agent/metadata" : {
      "get" : {
        "operationId" : "metadata",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentMetadataResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          }
        },
        "tags" : [ "agent-metadata-controller" ]
      }
    },
    "/agent/metadata/versions" : {
      "get" : {
        "operationId" : "versions",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CapabilityVersionsResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          }
        },
        "tags" : [ "agent-metadata-controller" ]
      }
    },
    "/agent/preflight" : {
      "post" : {
        "operationId" : "preflight",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PreflightRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PreflightResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "INVALID_PLAN"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "UNAUTHENTICATED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "NOT_PERMITTED"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "STALE_VERSION"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "NOT_FOUND, AMBIGUOUS_ENTITY, PRECONDITION_FAILED"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "NOT_IMPLEMENTED"
          }
        },
        "tags" : [ "preflight-controller" ]
      }
    },
    "/agent/session/capabilities" : {
      "get" : {
        "operationId" : "capabilities",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SessionCapabilitiesResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          }
        },
        "tags" : [ "agent-session-controller" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Local development",
    "url" : "http://localhost:8080"
  } ]
}